Added Sonobuoy Files and All Tests Are Added
Updated .gitignore

Added Sonobuoy Files and All Tests Are Added

I added a Dockerfile and other config files needed for running Sonobuoy.
I also updated the test file to remove references to kubectl that are no longer needed.

Initial Commit

Added files necessary for running tests: kube.yaml, requirements.txt, and readme.md
Added initial test (Note: This does not work yet)

Added changes to the test

This test is still not passing, but pushing it up to see if someone can take a look and figure out what I am doing wrong.

Changed a few more minor things, but the test is still not working

Changed a few more minor things, but the test is still not working

Newest changes to the test

Tests are still not running correctly, but this is the newest build for use by others.

Committing running test

The test is actually running on this one, but gets hung at the end, so I am going to try going back to the previous threading strategy. I am storing this for now to see if that will work, and if not to roll back.

PASSING TEST!!!!

O HAPPY DAY! The test is now running and passing. I need to add a few more, and to confirm that they will go to the preprocessed_logs_control_plane, but otherwise... O What A Wonderful Feeling!

Added All Happy Path Tests

I added all of the Happy Path Tests and ensured that they are passing consistently. Happy Path Tests are verifying the masked fields are being masked and that the Preprocessing Service is identifying if the log is from an RKE, RKE2, or k3s control plane or not.

Avoiding merge conflicts

I did not make changes to this file, so I am deleting it from my PR so that it does not cause merge conflicts with upstream changes.